% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SES_life.R
\name{SES_life}
\alias{SES_life}
\title{Historical Area-level Socioeconomic Deprivation for Edinburgh}
\usage{
SES_life(Lat, Long, Id, Year, Polygon)
}
\arguments{
\item{Lat}{has to be within Edinburgh boundary}

\item{Long}{has to be within Edinburgh boundary}

\item{Year}{has to be between 1926-2015}

\item{Polygon}{is the deprivation index shapefile downloaded from https://markcherrie.shinyapps.io/MMPgeodata/}
}
\description{
Derives an estimate of area-level socioeconomic deprivation for any Edinburgh location from 1926-2015.
}
\examples{
SES_life(55.9533, -3.1883, "ME1", 1948, Polygon)

}
\keyword{Edinburgh,}
\keyword{course,}
\keyword{deprivation}
\keyword{historical}
\keyword{index}
\keyword{life}
\keyword{of}

